(misc) add the concept of adapters
This adds Adapters to the broker that litens for messages on one
protocol and translates them to another

This can be used for many things, like:

  * receive registration message, republish onto NATS Streaming
  * receive replies, write to another system like ELK
  * receive messages from some other system and produce choria reqs

The first - republishing to NATS Streaming is included.

This also starts putting an instance of context.Context throughout
from the cmd system which can help with shutdown.  The federation
broker and server should be aware of this and shut down cleanly but
this adapter does it really nicely already