(misc) make the registration system an instance on a struct
It'll be great if the server component is entirely embeddable into
other code.  Imagine you want to build some specific IOT thing
and use the choria networking from a rpi but you want to compile
the agent into it that talks to GPIO.  This way you can compile
your own program that hosts a status of the choria server and
register your agent with it.

Or just the emmulator can run actual real instances rather than
emulated ones etc