(#57, #59) Create SimpleRPC compat layer and choria_util
This creates a basic MCollective RPC layer that can easily create
mcollective compatible RPC agents.  Right now its lacking in AAA and
request parameter handling but it's a good start

A ruby choria compatible choria_util agent is included that expose
internals in a matching way to the ruby one.