(#81) use prometheus as primary stats package
This undoes much of the work done so far, it replaces the metrics
library with the prometheus one as the main work horse for metrics.

The main problem is that while the metrics library has a export method
to prometheus its extremely lossy - as in, it does not support any
of the advanced prometheus features like labels etc.  This yields
the resulting metrics almost unusable by actual prometheus.

So now we have only /choria/prometheus HTTP endpoint, I anticpate
I'll write a prometheus -> simple exported that makes some kind
of JSON map or something of the data, if at all but at least its
possible and we're supporting the most feature rich protocol as
first class