(#120) initial provisioning agent
This adds a fairly simple agent with 2 actions - configure and restart

The configure agent takes a map[string]string as input and writes that
to the configuration file currently being used.  Care is taken to do this
via a tempfile and to ensure the resulting file is loadable as a choria
configuration file and does not again enable provisioning.

The restart action takes a splay setting which will be the time from now
that the server will restart - this will only happen if again the config
gets validated etc