(#163) limit the reconnect buffer space
The go nats library has a buffer used to store published messages
while the server reconnects, on reconnect it sends the buffer.

Once the buffer is full messages are not added to it - ie. only the
oldest stuff is sent.

This is pretty much useless to us in every scenario, no-one cares
for ancient registration stuff and replies - unless perhaps in an
async setup - has little value after much time.

Additionally this buffer is 30MB by default, imagine 20k nodes each
publishing 30MB of ancient rubbish on reconnect, this is not sensible
for us

So this sets the buffer to 10KB instead