(#176) intercept certain debug logs and elevate them
gnatsd logs mainly at debug irrespective of the usability of most
messages, this is a problem because I suspect I am seeing cases
where the single worker between brokers are being disconnected
as slow consumers.  The logs that would help identify that are
all debug and enabling debug on a large broker cluster is not possible

So this does a rather nasty intercept of those log messages and elevate
them to notice or error.