(#243) Call ruby agents via a shim
This creates the concept of a agent provider and impliments one
that can dispatch rpc requests to a shim that would call the ruby
agents.

In a future commit this provider will be configurable much like
additional agents are so that the mcorpc and other future ones
can be compiled in .. or not.

This also uses gomocks for the first time in tests