(#243) Create an agent provider framework
This allows new agent providers to be plugged into choria during
compile time.

An example agent provider is the one that wraps the oldm mcollective
ruby agents in a go shim so that they can be accessed via the existing
mcorpc system.

It adds build hooks and go generate hooks to include the ruby one
by default