(#302) hint when registration data changes
This sets a flag in the published data each time the registration data
changed.  The idea is that receivers like the stream replicator can
use this hint to replicate data out of their usual frequency ensuring
that the most fresh data is available.

We specifically ignore this flag at startup to avoid orchestrated
fleet wide restarts from overwhelming the consumers