Run broker/server as a user if specified on el6
This is more useful for the actual broker than the server, but will
honor specified user value from sysconfig override.

Note: Changing the value of the user without changing the value of the
pid file or logfile will probably result in a bad time.  There is also
no good way of changing these values unless the broker or server are
shut down first, since the init script depends on the pid file.