(misc) Remove trailing dot of SRV target
When using SRV records, the canonical hostname of the machine providing
the service is expected to end with a dot.

But if this dot is kept, TLS connection will fail because the hostname
does not match the Sever Name Indication (SNI):

```
org.eclipse.jetty.http.BadMessageException: 400: Host does not match SNI
```

Remove any trailing dot at the end of the target to fix this issue.