(misc) Setup Identity when running as a client only
Choria used to check if the user UID was different from 0 and in this
case was setting-up the Identity based on the user environment.  In
5b77ff95c7a990c6845008fac4272615d67fe6c1, we added support for windows
which does not provide such uid mecanism, and Identity was always set
according to the running user's environment.

This however cause trouble when running the Choria as a choria-server
service: by configuring Identity, the server does not use the expected
certificates.

Rework the way Identity is set to fix this issue: rely on
InitiatedByServer to determine if Choria is running as a server or not,
and only in this case setup Identity.