(#751) Improve project parsing
Previously we had InitiatedByServer to indicate server
specific behaviours and we overloaded that to assume all
system components are servers.  This is not true for the
broker.

Now we have ways to create a system configuration with
InitiatedBySystem set, the projects parsing and puppet
security system uses that to do the right thing - avoid
loading projects and use the system TLS when not configured

Signed-off-by: R.I.Pienaar <rip@devco.net>