(#1173) rework accounts and jetstream setup
This moves away from NATS typical account stores so we can
double down on our own JWT tokens.

Towards that we remove the account store and related files.
We create a choria account and move all connections into
this account, we also create a system account which can
be accessed using the new username/password configs - later
some jwt will give access.

Various tweaks to JetStream to enable it in the choria account

Signed-off-by: R.I.Pienaar <rip@devco.net>