(#1355) Support Curve 25519 secret sharing to protect the private key
This adds a EDCH Public key to the JWT response that the provisioner will
record and use to do Diffie-Hellman secret sharing.

The secret will be use to encryp the private key that the CA made and then
once sent to the provisioned server the provisioners Curve 25519 public key
is sent with the configuration request.

This way the key is encrypted over the wire using a single use DH shared secret

The jwt request essentially start a session and the following configure request
must deliver a 25519 Public key derived from teh servers.

After this any private key sent must be encrypted and the exchange must use
25519 secret sharing else the configure request is rejected.

25519 not used for any other part of the provision flow

Signed-off-by: R.I.Pienaar <rip@devco.net>