(#1377) remove some choria dependencies
This reworks Message and ConnectorMessage into interfaces

The goal is to have fewer places importing go-choria/choria
into their packages, the end goal is to allow security and
ddl lookups to make RPC requests.

At the moment this is entirely impossible since the interdependencies
cause import loops - rpc client imports fw, fw must use client.

So this is a start of a set of large and painful changes
to introduce interfaces for a number of core things

Signed-off-by: R.I.Pienaar <rip@devco.net>