(#1386) supports an optional sliding window timeout on broadcast
This adds CLI and configuration options that enables a sliding
window on discovery timeout.

Initially the timeout is as always but soon as the first reply
arrives the timeout is shrunk to 300ms, every subsequent reply
moves the 300ms timeout forward.

On my network this drops discovery by over 1 second and allows
for the initial slow start of replies.

It's opt-in so no default changes

Signed-off-by: R.I.Pienaar <rip@devco.net>