(maint) Rework Debian 11 packaging
The generic and bullseye packaging only differ in a build-dependency,
dh-systemd, which is not shipped anymore with Debian 11.

This package however appears to be a transition package which install
debhelper, the other build dependency we already have (since at least
Debian 8 [1]).

Removing the transition package from the list of build dependencies
should therefore be enough to build a Debian 11 package of choria.

Reverting #1376

[1]: https://packages.debian.org/stretch/dh-systemd