(#1388) Do not override dh_systemd_start/dh_systemd_enable
These scripts manage the lifecycle of the systemd unit files.  The
broker is configured to not be enabled and not be started, while the
server configuration has no special configuration, so it is enabled and
started duing installation and restarted during upgrades.

Because of the --no-start switch passed to dh_installinit when managing
the broker service, it is not automatically restarted during upgrades
(either dh_installinit manage the whole lifecycle, either it does not
manage it at all).