(#1241) Refactor DDL resolution
This refactors DDL resolution into providers with a cache, local file
and schema registry provider existing today.  Later maybe also a KV
provider to store schemas there.

Plugin doc, Completion, Req, Clients and various internals all updated
to transparently access the registry if configured to do so.

Removes an earlier attempt to do this that was not clean or abstracted
into provider

Signed-off-by: R.I.Pienaar <rip@devco.net>