(#1502) Upgrade KV, introduce formal Leader Election
This upgrades KV to the nats.go implementation and adds a
standard CHORIA_LEADER_ELECTION KV bucket that is to be used
for leader election.

We add a KV Based Leader Election provider that treats keys
in the bucket as elections and we create it as standard.

When the special provisioning account is enabled we import
one specific key into that account to facilitate LE for the
Choria Provisioner without exposing other elections into
that account

Signed-off-by: R.I.Pienaar <rip@devco.net>