(#1665) initial basic working app framework
This allows applications to be built that wraps RPC and KV
components.

Application definitions are documented using a schema and
are validated on loading, from the definition a CLI app is
built that can have nested sub commands and more.

Application definitions goes in files called x-app.yaml,
a symlink from x to choria would then load x-app.yaml and
construct the CLI when x is invoked.

An additional config file can be read called applications.yaml
that can be accessed via go templates in the definition in a few
places.

The config and definitions can live in ., ~/.config/choria/builder
or /etc/choria/builder.

Paths are using XDG, but its a super primitive implementation
for the moment, will be refined via issue 1624, thus paths
and file names are subject to change

Signed-off-by: R.I.Pienaar <rip@devco.net>