(#1713) allow system access with a jwt over unverified tls
Previously we insisted that system access needed verified TLS.
We still support that mode of access but now we extended the
client JWT with a claim allowing system access.

When a connection comes in over unverified TLS - however TLS
is required - and that connection has this claim set true
and that connection did the correct nonce signing and that
connection set system user/pass correctly then we join him
to the system account.

Signed-off-by: R.I.Pienaar <rip@devco.net>