(#1732) improve stdin json detection
The code that switches to flatfile in choria json
mode when there is no tty now reads the whole stdin
and detects ^{ and }$ before switching, if that is
not the case no switching is done, this should improve
the situation where choria is run in cron.

Also adjusts the order of checks for this mode making it
least significant.

Signed-off-by: R.I.Pienaar <rip@devco.net>