(#1755) add election evict
Also remove watch as its just not useful without the internal
knowledge of the statemachines.  It will show a leader
trying to maintain its leadership but without knowing
if its passing or failing you cant really understand
the message flow.  And getting replies is too costly
and also impossible in JWT world

Signed-off-by: R.I.Pienaar <rip@devco.net>