(#1798) support shutting down during provisioning
This allows provisioner to shut the instance down
cleanly so systemd will not restart it.

This supports the notion that in many environments
there are fully ephemeral nodes that could live for
a very short while and so little benefit from them
onboarding.

This can be use to just shut Choria down on them
and not waste further provisioning cycles on them

Signed-off-by: R.I.Pienaar <rip@devco.net>