(#1740) remove the concept of a cache from security providers
The cache was mainly used to find all privileged certificates
and then verify signatures against all known privileged certificates.

This was a useless effort since the privileged certificate was already
attached with the message.

So this remove the cache entirely and remove privileged checks
that was using the cache.

We pulled out the validation that was done during cache save into
a new security method and call that instead of save to gain the
validation benefits.

Follow up work is to write some integration tests to ensure the
security remains safe.

Signed-off-by: R.I.Pienaar <rip@devco.net>