(#1854) correct detection of empty filters
Filters were made to report empty when only 1 agent was set
as for appbuilder that made sense.

This however was a bad idea and had the effect that a request
with only agent filters set would be reported as empty to the
server when filtering incoming requests meaning all would reply
when they do not have the agent on them.

We fixed that and also setting the correct discovery method on
the choria req client

Signed-off-by: R.I.Pienaar <rip@devco.net>