(#1740) Adds context to parts of the protocol
Parts of the protocol call to remote services for signing
but did not previously have context, if things got stuck there
^c would simply hang until the fore quit kicks in

This adds context where required

Also goes a bit toward figuring out the various ways tokens
can be set, in the choria provider it now errors for some
and for others defaults to its own paths.  More work to do.

Signed-off-by: R.I.Pienaar <rip@devco.net>