(#1952) Fix fact filter parsing
With the addition of GJson to describe fact values the old fact parser is no
longer able to correctly identify facts vs operators in examples like
`'storage.#(name=="vda").size'`.

Here we rewrite the ParseFactFilterString function to no longer used complicated
regular expressions. We will now correctly identify GJson facts and parse the
filters correctly.

In addition we're adding a wider range of tests for ParseFactFilterString so
that we're covered if we were to add more fact filter functionality in the
future.