(#1987) Export SetBuildBasedOnJWT
resolver.setBuildBasedOnJWT is a method that other proftarget plugins often would like
to use.

Here we decouple the SetBuildBasedOnJWT method from the resolver struct and
export it as a function. To make the function usable outside of the resolver
struct we add a io.reader and build.info to the parameters.

Tests have also been updated to use the new function signature.