(#2133) Prepare the gossip service discovery system for nats 2.11
We previously attempted to build a service
discovery system in a KV bucket but we really
wanted reads to be a single API call.  So we
did what other systems like etcd suggest and
crammed it all in the key name.

This is restrictive and annoying, but in nats 2.11
we will be able to do a direct get on a kv bucket
that will return many machine values in a single
API call, this means we can now improve our registration
and add some adhoc data to it

So this adds annotations to the payload and simplify
the subjects being published moving the meat of it
into the message payload instead.

Signed-off-by: R.I.Pienaar <rip@devco.net>