(#2173) Load plugins configuration from the system directory
Some plugins need to load configuration that is managed system-wide by
Puppet even when the user is using their own configuration file.
Regardless of the configuration file loaded (system one or user one), we
want to load these plugins configuration from the system.

We previously inferred the plugins configuration directory relatively to
the rest of the configuration, but ignored it if it was located in the
user HOME directory.  This was not correct, and even caused more trouble
when the HOME environment variable was '/' as it is commonly on some
systems for services, because plugins configuration was never loaded on
these systems.

Fix this by detecting the location of the system configuration and
unconditionally loading plugins configuration from this directory.